split out files by types. more accessors
